# docs/swagger.yaml
basePath: /
definitions:
  commons.Metadata:
    properties:
      created_at:
        type: integer
      created_by:
        type: string
      deleted_at:
        type: integer
      deleted_by:
        type: string
      updated_at:
        type: integer
      updated_by:
        type: string
    type: object
  commons.Pagination:
    properties:
      page_number:
        type: integer
      page_size:
        type: integer
      total_item:
        type: integer
      total_pages:
        type: integer
    type: object
  commons.Response:
    properties:
      message:
        type: string
    type: object
  customerservice.Customer:
    properties:
      address:
        type: string
      country:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/commons.Metadata'
      name:
        type: string
      phone:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      social_name:
        type: string
      user_id:
        type: string
    type: object
  customerservice.ListCustomersResponse:
    properties:
      customers:
        items:
          $ref: '#/definitions/customerservice.Customer'
        type: array
      pagination:
        $ref: '#/definitions/commons.Pagination'
    type: object
  models.CreateCustomerRequest:
    properties:
      address:
        type: string
      country:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      social_name:
        type: string
    type: object
...
  models.Customer:
    properties:
      address:
        type: string
      country:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: string
      metadata:
        $ref: '#/definitions/valueobject.Metadata'
      name:
        type: string
      phone:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      social_name:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.UpdateCustomerRequest:
    properties:
      address:
        type: string
      country:
        type: string
      description:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      picture_url:
        type: string
      postal_code:
        type: string
      social_name:
        type: string
    type: object
...
  valueobject.Metadata:
    properties:
      created_at:
        type: integer
      created_by:
        type: string
      deleted_at:
        type: integer
      deleted_by:
        type: string
      updated_at:
        type: integer
      updated_by:
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: orkys.com@gmail.com
    name: API Support
    url: http://www.example.com/support
  description: This is the API documentation for the Orkys API Gateway
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://example.com/terms/
  title: Orkys API
  version: "1.0"
paths:
  /customers:
    get:
      consumes:
      - application/json
      description: Returns a list of customers with pagination.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Number of customers per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The customer list and pagination details
          schema:
            $ref: '#/definitions/customerservice.ListCustomersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns a list of customers.
      tags:
      - customer
    post:
      consumes:
      - application/json
      description: Creates a new customer with the provided details.
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: The customer object
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: The created customer object
          schema:
            $ref: '#/definitions/models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Creates a new customer.
      tags:
      - customer
  /customers/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a customer by ID.
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/commons.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Deletes a customer.
      tags:
      - customer
    get:
      consumes:
      - application/json
      description: Returns a customer by ID with all the details.
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The customer object
          schema:
            $ref: '#/definitions/models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns a customer by ID.
      tags:
      - customer
    patch:
      consumes:
      - application/json
      description: Updates a customer with the provided details.
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: The customer object
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: The updated customer object
          schema:
            $ref: '#/definitions/models.Customer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Updates a customer.
      tags:
      - customer
...
swagger: "2.0"